TravelRouting/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
altair==5.1.1
    # via streamlit
asttokens==2.4.0
    # via stack-data
attrs==23.1.0
    # via
    #   fiona
    #   jsonschema
    #   referencing
backcall==0.2.0
    # via ipython
blinker==1.6.2
    # via streamlit
branca==0.6.0
    # via
    #   folium
    #   streamlit-folium
cachetools==5.3.1
    # via streamlit
certifi==2023.7.22
    # via
    #   fiona
    #   pyproj
    #   requests
charset-normalizer==3.2.0
    # via requests
click==8.1.7
    # via
    #   click-plugins
    #   cligj
    #   fiona
    #   streamlit
click-plugins==1.1.1
    # via fiona
cligj==0.7.2
    # via fiona
cmake==3.27.5
    # via triton
comm==0.1.4
    # via ipykernel
concrete-python==2.0.0
    # via -r requirements.in
contourpy==1.1.1
    # via matplotlib
cycler==0.11.0
    # via matplotlib
debugpy==1.8.0
    # via ipykernel
decorator==5.1.1
    # via ipython
exceptiongroup==1.1.3
    # via ipython
executing==1.2.0
    # via stack-data
filelock==3.12.4
    # via
    #   torch
    #   triton
fiona==1.9.4.post1
    # via geopandas
folium==0.14.0
    # via
    #   -r requirements.in
    #   streamlit-folium
fonttools==4.42.1
    # via matplotlib
geopandas==0.14.0
    # via -r requirements.in
gitdb==4.0.10
    # via gitpython
gitpython==3.1.37
    # via streamlit
idna==3.4
    # via requests
importlib-metadata==6.8.0
    # via streamlit
ipykernel==6.25.2
    # via -r requirements.in
ipython==8.15.0
    # via ipykernel
jedi==0.19.0
    # via ipython
jinja2==3.1.2
    # via
    #   altair
    #   branca
    #   folium
    #   pydeck
    #   streamlit-folium
    #   torch
joblib==1.3.2
    # via scikit-learn
jsonschema==4.19.1
    # via altair
jsonschema-specifications==2023.7.1
    # via jsonschema
jupyter-client==8.3.1
    # via ipykernel
jupyter-core==5.3.1
    # via
    #   ipykernel
    #   jupyter-client
kiwisolver==1.4.5
    # via matplotlib
lit==16.0.6
    # via triton
mapclassify==2.6.0
    # via -r requirements.in
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.3
    # via jinja2
matplotlib==3.8.0
    # via -r requirements.in
matplotlib-inline==0.1.6
    # via
    #   ipykernel
    #   ipython
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
nest-asyncio==1.5.8
    # via ipykernel
networkx==3.1
    # via
    #   concrete-python
    #   mapclassify
    #   torch
numpy==1.26.0
    # via
    #   -r requirements.in
    #   altair
    #   concrete-python
    #   contourpy
    #   folium
    #   mapclassify
    #   matplotlib
    #   pandas
    #   pyarrow
    #   pydeck
    #   scikit-learn
    #   scipy
    #   shapely
    #   streamlit
nvidia-cublas-cu11==11.10.3.66
    # via
    #   nvidia-cudnn-cu11
    #   nvidia-cusolver-cu11
    #   torch
nvidia-cuda-cupti-cu11==11.7.101
    # via torch
nvidia-cuda-nvrtc-cu11==11.7.99
    # via torch
nvidia-cuda-runtime-cu11==11.7.99
    # via torch
nvidia-cudnn-cu11==8.5.0.96
    # via torch
nvidia-cufft-cu11==10.9.0.58
    # via torch
nvidia-curand-cu11==10.2.10.91
    # via torch
nvidia-cusolver-cu11==11.4.0.1
    # via torch
nvidia-cusparse-cu11==11.7.4.91
    # via torch
nvidia-nccl-cu11==2.14.3
    # via torch
nvidia-nvtx-cu11==11.7.91
    # via torch
packaging==23.1
    # via
    #   altair
    #   geopandas
    #   ipykernel
    #   matplotlib
    #   streamlit
pandas==2.1.1
    # via
    #   altair
    #   geopandas
    #   mapclassify
    #   streamlit
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==9.5.0
    # via
    #   matplotlib
    #   streamlit
platformdirs==3.10.0
    # via jupyter-core
prompt-toolkit==3.0.39
    # via ipython
protobuf==4.24.3
    # via streamlit
psutil==5.9.5
    # via ipykernel
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pyarrow==13.0.0
    # via streamlit
pydeck==0.8.1b0
    # via streamlit
pygments==2.16.1
    # via
    #   ipython
    #   rich
pyparsing==3.1.1
    # via matplotlib
pyproj==3.6.1
    # via geopandas
python-dateutil==2.8.2
    # via
    #   jupyter-client
    #   matplotlib
    #   pandas
    #   streamlit
pytz==2023.3.post1
    # via pandas
pyzmq==25.1.1
    # via
    #   ipykernel
    #   jupyter-client
referencing==0.30.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   folium
    #   streamlit
rich==13.5.3
    # via streamlit
rpds-py==0.10.3
    # via
    #   jsonschema
    #   referencing
scikit-learn==1.3.1
    # via mapclassify
scipy==1.11.2
    # via
    #   concrete-python
    #   mapclassify
    #   scikit-learn
shapely==2.0.1
    # via
    #   -r requirements.in
    #   geopandas
six==1.16.0
    # via
    #   asttokens
    #   fiona
    #   python-dateutil
smmap==5.0.1
    # via gitdb
stack-data==0.6.2
    # via ipython
streamlit==1.27.0
    # via
    #   -r requirements.in
    #   streamlit-folium
streamlit-folium==0.14.0
    # via -r requirements.in
sympy==1.12
    # via torch
tenacity==8.2.3
    # via streamlit
threadpoolctl==3.2.0
    # via scikit-learn
toml==0.10.2
    # via streamlit
toolz==0.12.0
    # via altair
torch==2.0.1
    # via
    #   concrete-python
    #   triton
tornado==6.3.3
    # via
    #   ipykernel
    #   jupyter-client
    #   streamlit
traitlets==5.10.0
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
triton==2.0.0
    # via torch
typing-extensions==4.8.0
    # via
    #   altair
    #   streamlit
    #   torch
tzdata==2023.3
    # via pandas
tzlocal==5.0.1
    # via streamlit
urllib3==2.0.5
    # via requests
validators==0.22.0
    # via streamlit
watchdog==3.0.0
    # via streamlit
wcwidth==0.2.6
    # via prompt-toolkit
wheel==0.41.2
    # via
    #   nvidia-cublas-cu11
    #   nvidia-cuda-cupti-cu11
    #   nvidia-cuda-runtime-cu11
    #   nvidia-curand-cu11
    #   nvidia-cusparse-cu11
    #   nvidia-nvtx-cu11
zipp==3.17.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
